move docs annotations up a level
