+++ backend/src/main/java/com/glossaar/backend/category/CategoryController.java
@@ -4,9 +4,9 @@ import com.glossaar.backend.category.dto.CategoryResponseDto;
 import com.glossaar.backend.category.dto.CategoryUpdateDto;
 import com.glossaar.backend.word.WordEntity;
 import com.glossaar.backend.word.WordService;
-import com.glossaar.backend.word.docs.BadRequestApiResponse;
-import com.glossaar.backend.word.docs.InternalServerErrorApiResponse;
-import com.glossaar.backend.word.docs.NotFoundApiResponse;
+import com.glossaar.backend.docs.BadRequestApiResponse;
+import com.glossaar.backend.docs.InternalServerErrorApiResponse;
+import com.glossaar.backend.docs.NotFoundApiResponse;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
