+++ backend/src/main/java/com/glossaar/backend/quiz/QuizController.java
@@ -3,9 +3,9 @@ package com.glossaar.backend.quiz;
 import com.glossaar.backend.quiz.dto.QuizBatchAnswerRequestDto;
 import com.glossaar.backend.quiz.dto.QuizQuestionResponseDto;
 import com.glossaar.backend.quiz.dto.QuizSubmitResponseDto;
-import com.glossaar.backend.word.docs.BadRequestApiResponse;
-import com.glossaar.backend.word.docs.InternalServerErrorApiResponse;
-import com.glossaar.backend.word.docs.NotFoundApiResponse;
+import com.glossaar.backend.docs.BadRequestApiResponse;
+import com.glossaar.backend.docs.InternalServerErrorApiResponse;
+import com.glossaar.backend.docs.NotFoundApiResponse;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.media.ArraySchema;
