+++ backend/src/main/java/com/glossaar/backend/word/WordController.java
@@ -4,16 +4,15 @@ import com.glossaar.backend.word.dto.CreateWordRequestDto;
 import com.glossaar.backend.word.dto.GetWordsResponseDto;
 import com.glossaar.backend.word.dto.UpdateWordRequestDto;
 import com.glossaar.backend.word.dto.WordResponseDto;
-import com.glossaar.backend.word.docs.BadRequestApiResponse;
-import com.glossaar.backend.word.docs.InternalServerErrorApiResponse;
-import com.glossaar.backend.word.docs.NotFoundApiResponse;
+import com.glossaar.backend.docs.BadRequestApiResponse;
+import com.glossaar.backend.docs.InternalServerErrorApiResponse;
+import com.glossaar.backend.docs.NotFoundApiResponse;
 import com.glossaar.backend.word.mapper.WordMapper;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
-import jakarta.transaction.Transactional;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
